Adding first support to forward queries (not yet)
I created a handler which role is to forward queries coming from clients while ignoring ones coming from another server. I should check how the queries information are stored to see if there is a need to make different methods for each type (GET/POST/etc.) or not. Also, there is an usage example in the class TableInsertion.