✨ Better forwarding condition
As of now, this configuration should work but it need to be adapted (what is consumed, what is produced, how we return the results of forwarded queries, etc.). Moreover, servers need to use a specific header in order to be recognized as a server by others when a forwarding happens. TableInsertion contains code example that isn't meant to stay there but stays as a reference for future implementations.