Forwarding queries (#4)
* Adding first support to forward queries (not yet)

I created a handler which role is to forward queries coming from clients while ignoring ones coming from another server. I should check how the queries information are stored to see if there is a need to make different methods for each type (GET/POST/etc.) or not. Also, there is an usage example in the class TableInsertion.

* Getting queries body and first try forwarding

* Smol change

* Forwarding is working (almost/maybe)

I need to test further to make adjustments but overall, the system should stay like this.

* Few improvements

- Cleared imports
- Simpler endpoint to test forwarding
- Another IP in the config
- Better check for local address to differentiate the different addresses accessible from client-side and not just the one in the configuration file.

* ✨ Better forwarding condition

As of now, this configuration should work but it need to be adapted (what is consumed, what is produced, how we return the results of forwarded queries, etc.). Moreover, servers need to use a specific header in order to be recognized as a server by others when a forwarding happens. TableInsertion contains code example that isn't meant to stay there but stays as a reference for future implementations.

* Adding new REST Client methods GET and PUT

Need to :
- Try-catch for connection errors
- Change GET to not have a body if not necessary
- How to parse queries responses
- Propagate the usage of the forwarding manager among all endpoints

---------

Co-authored-by: Faioa