Improvements (#21)
* Types support and changes to data structures

It works but still takes too much space (maybe even more than before, we need to make some improvements on that !

* ✨⚡️ Type Handling + Faster Insert

Now, exception are avoided correctly during insert which improves significantly the insert rate of our program. However, we still reach the OUT_OF_SPACE boundary very fast (which is strange because we allocated 10Go of RAM to the program, still need to research on that). We need to optimize the code and the objects we create more accurately. I don't think types can be more optimized than that but we can try on last resort (it doesn't affect the insert speed). Next step is to optimize the creation of objects to keep as much free space as we can.

* ✨ Trying to optimize...

* New project structure and new Main to allocate memory effectively

* ✨ Final Type Support

Types are now correctly parsed and the select works with it.

---------

Co-authored-by: Faioa <matthieu.pcf.dartois@gmail.com>