Changes for Select
Insertion is indeed a little bit faster but the select became of linear complexity. Thus, getting 3 columns for a single where condition took more that 1min30 (less than 23k results)... Now using a buffer that duplicates the values to get the O(1) complexity for the get by index. This buffer is not yet compressed as I encountered some problems when I tried (null values now get default values, which is not good) but we still earn a little bit of space from that. We can increase the gain by changing the compression algorithm to something stronger but slower, but let's focus on compressing the buffered values.