Tests to compress data array
Can compress the 3Mi lines up to 4200 Mb instead of 4800 Mb whithout any. We find a viable solution to not use the duplicated array or we try to improve this implementatione even further. I think speed might not be a problem since we don't use parallel execution yet but the memory space taken by the program is too large...