⚡️✨ It is now possible to insert multiple files in the same table
Moreover, the operations don't take that much memory anymore, increasing with the number of lines for the select.
However, the insert time is too long (now 28.5s for the first insert) and is increasing exponentially the more file we insert. To change this, we could use multi threading in the addRows method (Column class). We could search for a way to reduce the data taken by the select.
Also we don't use Brotli anymore since Snappy is much faster in decoding data (10s gain for the insert).