Indexes compression (#29)
* ⚡️ Faster (?) and Compressed (?) Insertion -> Need Tests

* Changes for Select

Insertion is indeed a little bit faster but the select became of linear complexity. Thus, getting 3 columns for a single where condition took more that 1min30 (less than 23k results)... Now using a buffer that duplicates the values to get the O(1) complexity for the get by index. This buffer is not yet compressed as I encountered some problems when I tried (null values now get default values, which is not good) but we still earn a little bit of space from that. We can increase the gain by changing the compression algorithm to something stronger but slower, but let's focus on compressing the buffered values.

* Tests to compress data array

Can compress the 3Mi lines up to 4200 Mb instead of 4800 Mb whithout any. We find a viable solution to not use the duplicated array or we try to improve this implementatione even further. I think speed might not be a problem since we don't use parallel execution yet but the memory space taken by the program is too large...

* Select returns the right rows but the values are incorrect

* ✅ Select now returns the right values

null value snow take more space, but the data is guaranteed to be correct.

* ⚡️ Different Changes

- Memory is now always freed so now the program takes way less heap memory (may need more advanced operation to run GarbageCollector when the heap is close to being full / creating a threshold to keep inserted data under this and still be able to perform operations)
- Removed unused dependency
- Closing streams after use

* ⚡️✨ It is now possible to insert multiple files in the same table

Moreover, the operations don't take that much memory anymore, increasing with the number of lines for the select.
However, the insert time is too long (now 28.5s for the first insert) and is increasing exponentially the more file we insert. To change this, we could use multi threading in the addRows method (Column class). We could search for a way to reduce the data taken by the select.
Also we don't use Brotli anymore since Snappy is much faster in decoding data (10s gain for the insert).

* ⚡️ Linear Insert Time

* 📝 New script for the presentation (TODO)

* On/Off compression in properties

* Faster insert and index not compressed

There is a NEED to compress indexes as it becomes 4 times bigger in memory without it

* ⚡ Indexes compression disabling and reel values compression disabling

* ⚡️ Changing how indexes are stored

We now don't use anymore presence byte since indexes cannot be null and are all unique values

* ✨ Deactivating indexation is now possible

* ✨ Dynamic indexation for columns

The threshold is configurable in the properties. Columns that contain too many unique values will not use the indexation. However, it is not possible to reactivate the indexation once it is deactivated.
Also changed some properties name.

* 🚧 Testing

* 🚧 Testing

* Small changes to wrap up this branch

---------

Co-authored-by: Faioa <matthieu.pcf.dartois@gmail.com>
Co-authored-by: Tsouli-Abde <93455423+Tsouli-Abde@users.noreply.github.com>