Removing unused controller
We now use responses with useful information on the state of the database so the debug controller loses a bit of its importance. We could remake it a lot better if we we focus on the data pertinence (one that is not already returned by the responses body like the state of all the tables etc.).